+++ crates/ruma-events/src/room/message/file.rs
@@ -1,5 +1,5 @@
 use js_int::UInt;
-use ruma_common::OwnedMxcUri;
+use ruma_common::MxcUri;
 use serde::{Deserialize, Serialize};
 
 use super::FormattedBody;
@@ -47,7 +47,7 @@ impl FileMessageEventContent {
     }
 
     /// Creates a new non-encrypted `FileMessageEventContent` with the given body and url.
-    pub fn plain(body: String, url: OwnedMxcUri) -> Self {
+    pub fn plain(body: String, url: MxcUri) -> Self {
         Self::new(body, MediaSource::Plain(url))
     }