+++ crates/ruma-events/src/room/message/video.rs
@@ -1,7 +1,7 @@
 use std::time::Duration;
 
 use js_int::UInt;
-use ruma_common::OwnedMxcUri;
+use ruma_common::MxcUri;
 #[cfg(feature = "unstable-msc2448")]
 use ruma_common::serde::Base64;
 use serde::{Deserialize, Serialize};
@@ -51,7 +51,7 @@ impl VideoMessageEventContent {
     }
 
     /// Creates a new non-encrypted `VideoMessageEventContent` with the given body and url.
-    pub fn plain(body: String, url: OwnedMxcUri) -> Self {
+    pub fn plain(body: String, url: MxcUri) -> Self {
         Self::new(body, MediaSource::Plain(url))
     }