+++ crates/ruma-events/tests/it/video.rs
@@ -7,7 +7,7 @@ use js_int::uint;
 use ruma_common::{
     MilliSecondsSinceUnixEpoch,
     canonical_json::assert_to_canonical_json_eq,
-    event_id, owned_mxc_uri,
+    event_id, mxc_uri,
     serde::{Base64, CanBeEmpty},
 };
 use ruma_events::{
@@ -25,10 +25,7 @@ use serde_json::{from_value as from_json_value, json};
 fn plain_content_serialization() {
     let event_content = VideoEventContent::with_plain_text(
         "Upload: my_video.webm",
-        FileContentBlock::plain(
-            owned_mxc_uri!("mxc://notareal.hs/abcdef"),
-            "my_video.webm".to_owned(),
-        ),
+        FileContentBlock::plain(mxc_uri!("mxc://notareal.hs/abcdef"), "my_video.webm".to_owned()),
     );
 
     assert_to_canonical_json_eq!(
@@ -50,7 +47,7 @@ fn encrypted_content_serialization() {
     let event_content = VideoEventContent::with_plain_text(
         "Upload: my_video.webm",
         FileContentBlock::encrypted(
-            owned_mxc_uri!("mxc://notareal.hs/abcdef"),
+            mxc_uri!("mxc://notareal.hs/abcdef"),
             "my_video.webm".to_owned(),
             EncryptedContent::new(
                 V2EncryptedFileInfo::new(
@@ -100,7 +97,7 @@ fn event_serialization() {
             "Upload: <strong>my_lava_lamp.webm</strong>",
         ),
         FileContentBlock::plain(
-            owned_mxc_uri!("mxc://notareal.hs/abcdef"),
+            mxc_uri!("mxc://notareal.hs/abcdef"),
             "my_lava_lamp.webm".to_owned(),
         ),
     );
@@ -112,7 +109,7 @@ fn event_serialization() {
     content.video_details = Some(video_details);
     let mut thumbnail = Thumbnail::new(
         ThumbnailFileContentBlock::plain(
-            owned_mxc_uri!("mxc://notareal.hs/thumbnail"),
+            mxc_uri!("mxc://notareal.hs/thumbnail"),
             "image/jpeg".to_owned(),
         ),
         ThumbnailImageDetailsContentBlock::new(uint!(560), uint!(480)),